Replace static values file
