+++ src/lib/client/Static/DisplayValues.ts
@@ -1,6 +1,16 @@
 export const colorStone200 = '#E7E5E4';
 export const colorStone700 = '#44403C';
 
+export const pageTitles = {
+	home: 'ホーム',
+	books: 'ライブラリ',
+	wish: '読みたい本',
+	reading: '読んでいる本',
+	complete: '読み終わった本',
+	search: '書籍検索',
+	shelf: '本棚'
+};
+
 export const emptyMessages = {
 	default: '本が登録されていません。<br>検索メニューから書籍を登録してください！',
 	wish: '読みたい本が登録されていません。<br>右上の追加ボタンから本を検索して登録してください！',
