+++ src/lib/components/menu/SideMenu.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
 	import { goto } from '$app/navigation';
 	import { page } from '$app/stores';
-	import { mainMenuItems } from '$lib/client/UI/Shared/DisplayData';
-	import { colorStone200 } from '$lib/client/UI/Shared/StaticValues';
+	import { mainMenuItems } from '$lib/client/Static/MenuItems';
+	import { colorStone200 } from '$lib/client/Static/DisplayValues';
 	import { firebaseAuth } from '$lib/firebase.client';
 	import Icon from '@iconify/svelte';
 	import { signOut } from 'firebase/auth';
