+++ src/routes/(app)/books/complete/+page.svelte
@@ -1,8 +1,7 @@
 <script lang="ts">
 	import { page } from '$app/stores';
 	import { setPathNameContext } from '$lib/client/Helpers/Svelte/ContextAPI';
-	import { pageTitles } from '$lib/client/UI/Shared/DisplayData';
-	import { emptyMessages } from '$lib/client/UI/Shared/StaticValues';
+	import { emptyMessages, pageTitles } from '$lib/client/Static/DisplayValues';
 	import type { selectFilterItem, toggleFilterItem } from '$lib/customTypes';
 	import CompleteBook from '$lib/icons/CompleteBook.svelte';
 	import BooksMainContent from '../BooksMainContent.svelte';
